+++ src/types/shemas/item.ts
@@ -1,0 +1,11 @@
+/** 아이템 정보 */
+export interface IItem {
+  /** 아이템 식별자 */
+  id: number;
+  /** 아이템 이름 */
+  name: string;
+  /** 아이템 카테고리 */
+  category: string;
+  /** 아이템 가격 */
+  price: number;
+}
