+++ src/types/shemas/user.ts
@@ -1,0 +1,9 @@
+/** 유저 정보 */
+export interface IUser {
+  /** 유저 식별자 */
+  id: number;
+  /** 유저 이름 */
+  name: string;
+  /** 유저 이메일 */
+  email: string;
+}
