+++ src/types/theme.ts
@@ -1,0 +1,6 @@
+/**
+ * 테마 타입
+ * 1. `light`: 밝은 테마
+ * 2. `dark`: 어두운 테마
+ */
+export type TTheme = "light" | "dark";
