+++ vite.config.ts
@@ -1,10 +1,14 @@
 import { defineConfig as defineTestConfig, mergeConfig } from "vitest/config";
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react-swc";
+import path from "path";
 
 export default mergeConfig(
   defineConfig({
     plugins: [react()],
+    resolve: {
+      alias: [{ find: "#src", replacement: path.resolve(__dirname, "src") }],
+    },
   }),
   defineTestConfig({
     test: {
